# src/main/resources/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/diaries/{diaryId}":{"get":{"tags":["diary-controller"],"operationId":"getDiary","parameters":[{"name":"diaryId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoDiaryResDto"}}}}}},"put":{"tags":["diary-controller"],"operationId":"updateDiary","parameters":[{"name":"diaryId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiaryUpdateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoDiaryResDto"}}}}}},"delete":{"tags":["diary-controller"],"operationId":"deleteDiary","parameters":[{"name":"diaryId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoVoid"}}}}}}},"/diaries":{"post":{"tags":["diary-controller"],"operationId":"createDiary","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"content","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoDiaryResDto"}}}}}}},"/dialogue/chat":{"post":{"tags":["dialogue-chat-controller"],"operationId":"streamChatResponse","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DialogueChatReqDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"text/event-stream":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/chat/stream":{"post":{"tags":["chat-controller"],"operationId":"streamDiaryFromConversation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageReqDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/x-ndjson":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatMessageResDto"}}}}}}}},"/diaries/user/{userId}":{"get":{"tags":["diary-controller"],"operationId":"getDiaries","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoDiaryResDto"}}}}}}},"/dialogue/{userId}/chat-list":{"get":{"tags":["dialogue-chat-controller"],"operationId":"getChatList","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoDialogueChatSessionResDto"}}}}}}},"/dialogue/{userId}/chat-list/{sessionId}":{"get":{"tags":["dialogue-chat-controller"],"operationId":"getChatDetail","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"sessionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"cursor","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoDialogueChatMessageResDto"}}}}}}}},"components":{"schemas":{"DiaryUpdateDto":{"type":"object","properties":{"content":{"type":"string"}}},"DiaryResDto":{"type":"object","properties":{"diaryId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"content":{"type":"string"},"createdAt":{"type":"string"}}},"MetaResponseDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"code":{"type":"string"},"message":{"type":"string"}}},"ResponseDtoDiaryResDto":{"type":"object","properties":{"meta":{"$ref":"#/components/schemas/MetaResponseDto"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DiaryResDto"}}}},"DialogueChatReqDto":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"gender":{"type":"string","enum":["MAN","WOMAN"]},"hobby":{"type":"array","items":{"type":"string","enum":["PIANO","GUITAR","DRUMS","VIOLIN","SINGING","COMPOSITION","DJ","UKULELE","BASS","FLUTE","SAXOPHONE","CELLO","TRUMPET","HARP","HARMONICA","KARAOKE","BAND","ORCHESTRA","RAP","HIPHOP","POP","KPOP","JPOP","CLASSICAL","CHOIR","JAZZ","ROCK","BLUES","METAL","REGGAE","COUNTRY","FOLK","EDM","RNB","SYNTH","BEATBOX","IMPROV","PRODUCTION","MIXING","THEORY","LIVE","RECORDCOLLECTING","COVER","REMIX","SOUNDDESIGN","BOARDGAMES","CARDGAMES","PUZZLE","CHESS","BADUK","SUDOKU","CROSSWORD","RPG","ESCAPEROOM","ESPORTS","MOBILEGAMES","CONSOLEGAMING","PCGAMES","ARCADE","QUIZGAMES","STRATEGYGAMES","SIMULATION","MAHJONG","POKER","SPEEDCUBING","MOVIEWATCHING","CLASSICFILMS","INDIEFILMS","DOCUMENTARY","FILMFESTIVAL","BINGEWATCHING","STREAMING","DRAMASERIES","KDRAMA","SITCOM","REALITYTV","TALKSHOW","ANIMEWATCHING","CARTOONWATCHING","WEBTOON","COSPLAY","STANDUPCOMEDY","COMEDY","THEATERGOING","MUSICALS","JOGGING","YOGA","PILATES","HIKING","CYCLING","SWIMMING","SOCCER","BASKETBALL","TENNIS","VOLLEYBALL","BASEBALL","GOLF","CLIMBING","SKATEBOARDING","SNOWBOARDING","SKIING","BOXING","MARTIALARTS","FENCING","ARCHERY","BILLIARDS","BOWLING","TAEKWONDO","DANCE","BADMINTON","TABLETENNIS","HANDBALL","ULTIMATEFRISBEE","ROWING","CANOEING","KAYAKING","PADDLEBOARDING","PARKOUR","SLACKLINING","WEIGHTLIFTING","CROSSFIT","SURFING","TRAMPOLINE","SCUBADIVING","EQUESTRIAN","WRESTLING","JUDO","KARATE","TAICHI","NOVEL","ESSAY","NONFICTION","POETRY","HISTORY","PHILOSOPHY","SCIENCE","MAGAZINE","COMIC","JOURNAL","FLASHCARD","SPEEDREADING","AUDIOBOOK","CLUB","DISCUSSION","FOREIGNLANGUAGE","ONLINECOURSE","ANNOTATION","TRANSCRIPTION","CAT","DOG","BIRD","RABBIT","FISH","INSECT","GARDENING","PLANTGROWING","BONSAI","TERRARIUM","FISHKEEPING","BEEKEEPING","MUSHROOMFORAGING","FOSSILCOLLECTION","STARGAZING","CAMPING","NATUREWALK","FORAGING","URBANFARMING","WILDLIFEWATCHING","TREEPLANTING","FORESTBATHING","FARM","ZOOVISITING","DRAWING","PAINTING","WATERCOLOR","OILPAINTING","CALLIGRAPHY","SCULPTURE","KNITTING","NAILART","SEWING","EMBROIDERY","CANDLEMAKING","SOAPMAKING","SCRAPBOOKING","MODELBUILDING","LEGO","WOODWORKING","METALCRAFT","POTTERY","TRAVEL","BACKPACKING","URBANEXPLORATION","PHOTOGRAPHY","FILMPHOTOGRAPHY","DRONEFLYING","THREEDPRINTING","ARDUINO","CODING","PROGRAMMING","AI","ROBOTICS","JOURNALING","BLOGGING","CREATIVEWRITING","POETRYWRITING","VOLUNTEER","LANGUAGEEXCHANGE","TRANSLATION","PODCASTLISTENING","MAGIC"]}},"todayDiary":{"type":"string"},"userInput":{"type":"string"}}},"ChatMessageReqDto":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"conversation":{"type":"string"},"currentDiary":{"type":"string"}}},"ChatMessageResDto":{"type":"object","properties":{"type":{"type":"string"},"message":{"type":"string"}}},"DialogueChatSessionResDto":{"type":"object","properties":{"sessionId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"lastMessageInfoDto":{"$ref":"#/components/schemas/LastMessageInfoDto"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"LastMessageInfoDto":{"type":"object","properties":{"messageId":{"type":"integer","format":"int64"},"chatRole":{"type":"string","enum":["ASSISTANT","USER"]},"message":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ResponseDtoDialogueChatSessionResDto":{"type":"object","properties":{"meta":{"$ref":"#/components/schemas/MetaResponseDto"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DialogueChatSessionResDto"}}}},"DialogueChatMessageResDto":{"type":"object","properties":{"messages":{"type":"array","items":{"$ref":"#/components/schemas/MessageInfoDto"}},"nextCursor":{"type":"integer","format":"int64"},"hasNextPage":{"type":"boolean"}}},"MessageInfoDto":{"type":"object","properties":{"messageId":{"type":"integer","format":"int64"},"sessionId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"chatRole":{"type":"string","enum":["ASSISTANT","USER"]},"message":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ResponseDtoDialogueChatMessageResDto":{"type":"object","properties":{"meta":{"$ref":"#/components/schemas/MetaResponseDto"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DialogueChatMessageResDto"}}}},"ResponseDtoVoid":{"type":"object","properties":{"meta":{"$ref":"#/components/schemas/MetaResponseDto"},"data":{"type":"array","items":{"type":"object"}}}}}}}